src/Backend/OCR_Backend/requirements.txt
Django==4.1.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
django-cors-headers==3.14.0
datasets==2.12.0
more_itertools==9.1.0
rpyc==5.3.1
datasets==2.12.0
rpyc==5.3.1
Pillow==9.4.0
numpy==1.24.2
deskew==1.4.3
jdeskew==0.2.1
ultralytics==8.0.115
gunicorn==20.1.0
boto3==1.26.151
